SearchRobot/pyproject.toml
[project]
name = "searchrobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "lxml>=6.0.2",
    "pybind11>=3.0.1",
    "pymongo>=4.16.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rq>=2.6.1",
    "rq-scheduler>=0.14.0",
    "scrapy>=2.14.1",
    "scrapy-redis>=0.9.1",
    "scrapy-user-agents>=0.1.1",
    "uvicorn>=0.40.0",
]
